ARTIF-725 ARTIF-726 Use new 'expandedFromArchivePath' field (rather than custom property), auto-deleted expanded artifacts if an archive is deleted

+++ test/src/test/java/org/artificer/test/AbstractIntegrationTest.java
@@ -69,7 +69,7 @@ public abstract class AbstractIntegrationTest {
             QueryResultSet results = client.query("/s-ramp", 0, 10000, "name", true);
             for (ArtifactSummary summary : results) {
                 String uuid = summary.getUuid().replace("urn:uuid:", "");
-                if (!summary.isDerived()) {
+                if (!summary.isDerived() && !summary.isExpandedFromArchive()) {
                     client.deleteArtifact(uuid, summary.getArtifactType(), true);
                 }
             }
